pyproject.toml
[project]
name = "genai-tools"
version = "0.1.0"
description = "A collection of generative AI tools including text-to-speech, document processing, and research utilities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.72.0",
    "beautifulsoup4>=4.13.5",
    "click>=8.2.1",
    "ebooklib>=0.19",
    "feedparser>=6.0.12",
    "html2text>=2025.4.15",
    "jupyter>=1.1.1",
    "litellm>=1.76.1",
    "llm>=0.19.0",
    "llm-gemini>=0.6.0",
    "matplotlib>=3.10.6",
    "nbconvert>=7.16.6",
    "openai>=1.104.2",
    "pydub>=0.25.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "torch>=2.8.0",
    "transformers>=4.51.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pylint>=3.3.0",
    "pre-commit>=4.0.0",
]

[tool.uv.sources]
vibevoice = { git = "https://github.com/microsoft/VibeVoice.git" }

[tool.ruff]
target-version = "py311"
line-length = 88
extend-exclude = [
    "assets/",
    ".venv/",
    "__pycache__/",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
    "C4", # flake8-comprehensions
    "B",  # flake8-bugbear
    "A",  # flake8-builtins
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "A002",  # argument shadows built-in
]

[tool.ruff.lint.per-file-ignores]
"src/__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[project.scripts]
artomd = "tools.arxiv_to_md.cli:main"

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pylint>=3.3.8",
    "ruff>=0.12.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]
